+++ backend-rust/src/indexer.rs
@@ -626,6 +626,7 @@ impl ProcessEvent for BlockProcessor {
             }
             block.special_transaction_outcomes.save(&mut tx).await?;
             block.baker_unmark_suspended.save(&mut tx).await?;
+
             out.push_str(format!("\n- {}:{}", block.height, block.hash).as_str());
         }
         process_release_schedules(new_context.last_block_slot_time, &mut tx)
@@ -876,6 +877,7 @@ impl PreparedBlock {
         )
         .await?;
         let baker_unmark_suspended = PreparedUnmarkPrimedForSuspension::prepare(data)?;
+
         Ok(Self {
             hash,
             height,
@@ -938,18 +940,26 @@ impl PreparedBlock {
         }
 
         sqlx::query!(
-            "INSERT INTO blocks
-  (height, hash, slot_time, block_time, baker_id, total_amount, total_staked, cumulative_num_txs)
-SELECT * FROM UNNEST(
-  $1::BIGINT[],
-  $2::TEXT[],
-  $3::TIMESTAMPTZ[],
-  $4::BIGINT[],
-  $5::BIGINT[],
-  $6::BIGINT[],
-  $7::BIGINT[],
-  $8::BIGINT[]
-);",
+            "INSERT INTO blocks (
+                height, 
+                hash, 
+                slot_time, 
+                block_time, 
+                baker_id, 
+                total_amount, 
+                total_staked, 
+                cumulative_num_txs
+            )
+            SELECT * FROM UNNEST(
+                $1::BIGINT[],
+                $2::TEXT[],
+                $3::TIMESTAMPTZ[],
+                $4::BIGINT[],
+                $5::BIGINT[],
+                $6::BIGINT[],
+                $7::BIGINT[],
+                $8::BIGINT[]
+            );",
             &heights,
             &hashes,
             &slot_times,
@@ -4589,13 +4599,13 @@ struct PreparedValidatorSuspension {
 /// Represents a payday block, its payday commission
 /// rates, and the associated block height.
 struct PreparedPayDayBlock {
-    block_height:            i64,
+    block_height:                 i64,
     /// Represents the payday baker pool commission rates captured from
     /// the `get_bakers_reward_period` node endpoint.
-    payday_commission_rates: PreparedPaydayCommissionRates,
+    payday_commission_rates:      PreparedPaydayCommissionRates,
     /// Represents the payday lottery power updates for bakers captured from
     /// the `get_election_info` node endpoint.
-    bakers_lottery_powers:   PreparedPaydayLotteryPowers,
+    payday_bakers_lottery_powers: PreparedPaydayLotteryPowers,
 }
 
 impl PreparedPayDayBlock {
@@ -4625,12 +4635,13 @@ impl PreparedPayDayBlock {
             .get_election_info(BlockIdentifier::AbsoluteHeight(block_height))
             .await?
             .response;
-        let bakers_lottery_powers = PreparedPaydayLotteryPowers::prepare(election_info.bakers)?;
+        let payday_bakers_lottery_powers =
+            PreparedPaydayLotteryPowers::prepare(election_info.bakers)?;
 
         Ok(Self {
             block_height: block_height.height.try_into()?,
             payday_commission_rates,
-            bakers_lottery_powers,
+            payday_bakers_lottery_powers,
         })
     }
 
@@ -4642,7 +4653,7 @@ impl PreparedPayDayBlock {
         self.payday_commission_rates.save(tx).await?;
 
         // Save the lottery_powers to the database.
-        self.bakers_lottery_powers.save(tx).await?;
+        self.payday_bakers_lottery_powers.save(tx).await?;
 
         sqlx::query!(
             "UPDATE current_chain_parameters
@@ -4734,15 +4745,23 @@ impl PreparedPaydayCommissionRates {
 struct PreparedPaydayLotteryPowers {
     baker_ids:             Vec<i64>,
     bakers_lottery_powers: Vec<BigDecimal>,
+    ranks:                 Vec<i64>,
 }
 
 impl PreparedPaydayLotteryPowers {
-    fn prepare(bakers: Vec<BirkBaker>) -> anyhow::Result<Self> {
+    fn prepare(mut bakers: Vec<BirkBaker>) -> anyhow::Result<Self> {
         let capacity = bakers.len();
         let mut baker_ids: Vec<i64> = Vec::with_capacity(capacity);
         let mut bakers_lottery_powers: Vec<BigDecimal> = Vec::with_capacity(capacity);
+        let mut ranks: Vec<i64> = Vec::with_capacity(capacity);
 
-        for baker in bakers.iter() {
+        // Sort bakers by lottery power. The baker with the highest lottery power comes
+        // first in the vector and gets rank 1.
+        bakers.sort_by(|self_baker, other_baker| {
+            self_baker.baker_lottery_power.total_cmp(&other_baker.baker_lottery_power).reverse()
+        });
+
+        for (rank, baker) in bakers.iter().enumerate() {
             baker_ids.push(i64::try_from(baker.baker_id.id.index)?);
             bakers_lottery_powers.push(
                 BigDecimal::from_f64(baker.baker_lottery_power)
@@ -4752,11 +4771,13 @@ impl PreparedPaydayLotteryPowers {
                     )
                     .map_err(RPCError::ParseError)?,
             );
+            ranks.push((rank + 1) as i64);
         }
 
         Ok(Self {
             baker_ids,
             bakers_lottery_powers,
+            ranks,
         })
     }
 
@@ -4774,13 +4795,16 @@ impl PreparedPaydayLotteryPowers {
         sqlx::query!(
             "INSERT INTO bakers_payday_lottery_powers (
                 id,
-                payday_lottery_power
+                payday_lottery_power,
+                payday_ranking_by_lottery_powers
             )
             SELECT
                 UNNEST($1::BIGINT[]) AS id,
-                UNNEST($2::NUMERIC[]) AS payday_lottery_power",
+                UNNEST($2::NUMERIC[]) AS payday_lottery_power,
+                UNNEST($3::BIGINT[]) AS payday_ranking_by_lottery_powers",
             &self.baker_ids,
-            &self.bakers_lottery_powers
+            &self.bakers_lottery_powers,
+            &self.ranks
         )
         .execute(tx.as_mut())
         .await?;