+++ backend-rust/src/migrations.rs
@@ -197,14 +197,16 @@ pub enum SchemaVersion {
     StakedPoolSizeConstraint,
     #[display("0010:Add delegated stake cap")]
     DelegatedStakeCap,
+    #[display("0011:RankingByLotteryPower")]
+    RankingByLotteryPower,
 }
 impl SchemaVersion {
     /// The minimum supported database schema version for the API.
     /// Fails at startup if any breaking database schema versions have been
     /// introduced since this version.
-    pub const API_SUPPORTED_SCHEMA_VERSION: SchemaVersion = SchemaVersion::DelegatedStakeCap;
+    pub const API_SUPPORTED_SCHEMA_VERSION: SchemaVersion = SchemaVersion::RankingByLotteryPower;
     /// The latest known version of the schema.
-    const LATEST: SchemaVersion = SchemaVersion::DelegatedStakeCap;
+    const LATEST: SchemaVersion = SchemaVersion::RankingByLotteryPower;
 
     /// Parse version number into a database schema version.
     /// None if the version is unknown.
@@ -230,6 +232,7 @@ impl SchemaVersion {
             SchemaVersion::AccountBaseAddress => false,
             SchemaVersion::StakedPoolSizeConstraint => false,
             SchemaVersion::DelegatedStakeCap => false,
+            SchemaVersion::RankingByLotteryPower => false,
         }
     }
 
@@ -299,7 +302,15 @@ impl SchemaVersion {
                 )
                 .await?
             }
-            SchemaVersion::DelegatedStakeCap => unimplemented!(
+            SchemaVersion::DelegatedStakeCap => {
+                tx.as_mut()
+                    .execute(sqlx::raw_sql(include_str!(
+                        "./migrations/m0011-ranking-by-lottery-power.sql"
+                    )))
+                    .await?;
+                SchemaVersion::RankingByLotteryPower
+            }
+            SchemaVersion::RankingByLotteryPower => unimplemented!(
                 "No migration implemented for database schema version {}",
                 self.as_i64()
             ),